requirements.txt
hydra-core==1.3.1
lexicalrichness==0.3.1
nltk==3.8
numpy==1.24.1
pandas==1.5.2
pyenchant==3.2.2
scikit_learn==1.2.0
torch==1.13.1
tqdm==4.64.1
transformers==4.25.1